feat: 픽잇 식당 모두 소거 시 종료 기능 구현
