+++ frontend/src/domains/matchResult/hooks/usePickeatEndCheck.ts
@@ -1,23 +1,36 @@
 import { pickeat, PickeatStateResponse } from '@apis/pickeat';
+import { restaurants } from '@apis/restaurants';
 
 import { usePolling } from '@hooks/usePolling';
 
 import { generateRouterPath, ROUTE_PATH } from '@routes/routePath';
 
+import { useState } from 'react';
 import { useNavigate } from 'react-router';
 
 export function usePickeatStateChecker(pickeatCode: string) {
+  const [hasRestaurants, setHasRestaurants] = useState(true);
   const navigate = useNavigate();
 
   const fetchPickeatState = async (): Promise<PickeatStateResponse | null> => {
     return await pickeat.getPickeatState(pickeatCode);
   };
 
+  const fetchRestaurantList = async () => {
+    return await restaurants.get(pickeatCode, {
+      isExcluded: 'false',
+    });
+  };
+
   usePolling<PickeatStateResponse | null>(fetchPickeatState, {
-    onData: data => {
-      if (data?.isActive === false) {
-        navigate(generateRouterPath.matchResult(pickeatCode));
+    onData: async data => {
+      if (data?.isActive === true) return;
+      const restaurantList = await fetchRestaurantList();
+      if (restaurantList.length === 0) {
+        setHasRestaurants(false);
+        return;
       }
+      navigate(generateRouterPath.matchResult(pickeatCode));
     },
     errorHandler: error => {
       if (error.message === 'PICKEAT_NOT_FOUND') {
@@ -30,4 +43,6 @@ export function usePickeatStateChecker(pickeatCode: string) {
     interval: 3000,
     immediate: true,
   });
+
+  return { hasRestaurants };
 }
